+++ packages/antd-management-fast-design-playground/src/components/displayer/RemarkEditor/index.jsx
@@ -31,6 +31,7 @@ class RemarkEditor extends BaseComponent {
       dataTag: '',
       dataStorage: [],
       dataBackup: [],
+      color: '',
     };
   }